moved components into their relevant directories
