+++ app/dashboard/page.tsx
@@ -1,5 +1,5 @@
 'use client';
-import RosterForm from '../ui/roster-form';
+import RosterForm from '../start/components/roster-form';
 
 const Page = () => {
   return (
