+++ app/start/components/assign-details.tsx
