+++ app/start/components/grid-cell.tsx
