+++ app/start/components/progress-bar-item.tsx
