+++ app/start/components/roster-form.tsx
@@ -13,13 +13,13 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import { useForm } from 'react-hook-form';
 import useMultiStepForm from '@/app/lib/useMultiStepForm';
-import ProgressBar from '@/app/ui/multistep-form/progress-bar';
+import ProgressBar from '@/app/start/components/progress-bar';
 
 // import types, schemas and the other two components
-import { Day, formSchema } from '../lib/formSchemas';
+import { Day, formSchema } from '../../lib/formSchemas';
 import WorkDetails from './work-details';
 import EmployeeDetails from './employee-details';
-import GridSelector from '@/app/ui/grid-selector/grid-selector';
+import GridSelector from '@/app/start/components/grid-selector';
 
 const StartForm = () => {
   const days: Day[] = [
