+++ app/start/page.tsx
@@ -1,0 +1,8 @@
+export default function Page()
+{
+    return(
+        <>
+            
+        </>
+    )
+}
